src/requirements-testing.txt
pytest==2.9.0
pytest-django==2.9.1
flake8==2.5.4
factory-boy==2.6.1
